Add song editor with drag-and-drop reordering
- Create separate Player and Editor views with routing
- Extract WebAudioSynth class to synth.ts for reuse
- Extract utility functions to utils.ts (getNotesFromAnchor, generateRandomSong)
- Add song editor at /editor route with:
  - Add notes by octave (3-6) and chromatic scale
  - Remove individual notes or clear all
  - Drag-and-drop reordering (desktop and mobile)
  - Visual note previews on click
  - Done/Cancel buttons to save or discard changes
- Update Player with Edit/Create Song button
- Update README with ponrappa.pages.dev example URL
- Update page title to "PonRappa"
- Add Japanese name meaning to app and README
- Improve button styling for consistency
- Handle empty hash parameter (#m=) as no song

Mobile drag-and-drop uses long-press (200ms) to activate dragging with visual feedback.

via [Happy](https://happy.engineering)
